requirements.txt
# FKS Monitor Service Dependencies

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
aiohttp>=3.9.0
httpx>=0.25.0

# Configuration
pyyaml>=6.0.1

# Prometheus
prometheus-client>=0.19.0

# Google AI API (for test analysis)
google-generativeai>=0.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.1.0
mypy>=1.7.0

